refactor command and add vela generate-doc for cli
