+++ documentation/cli/vela.md
@@ -1,10 +1,10 @@
 ## vela
 
-✈️  A Micro App Platform for Kubernetes.
+
 
 ### Synopsis
 
-✈️  A Micro App Platform for Kubernetes.
+
 
 ```
 vela [flags]
@@ -25,14 +25,17 @@ vela [flags]
 * [vela completion](vela_completion.md)	 - Output shell completion code for the specified shell (bash or zsh)
 * [vela dashboard](vela_dashboard.md)	 - Setup API Server and launch Dashboard
 * [vela env](vela_env.md)	 - Manage application environments
+* [vela init](vela_init.md)	 - Init an OAM Application
 * [vela install](vela_install.md)	 - Initialize vela on both client and server
 * [vela logs](vela_logs.md)	 - Tail pods logs of an application
-* [vela manualscaler](vela_manualscaler.md)	 - Attach manualscaler trait to an app
-* [vela rollout](vela_rollout.md)	 - Attach rollout trait to an app
+* [vela metrics](vela_metrics.md)	 - Attach metrics trait to an app
+* [vela route](vela_route.md)	 - Attach route trait to an app
 * [vela scale](vela_scale.md)	 - Attach scale trait to an app
 * [vela system](vela_system.md)	 - system management utilities
+* [vela template](vela_template.md)	 - Manage templates
 * [vela traits](vela_traits.md)	 - List traits
+* [vela up](vela_up.md)	 - Apply an appfile
 * [vela version](vela_version.md)	 - Prints out build version information
 * [vela workloads](vela_workloads.md)	 - List workloads
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
