+++ documentation/cli/vela_app.md
@@ -33,4 +33,4 @@ vela app <command>
 * [vela app show](vela_app_show.md)	 - get details of your app
 * [vela app status](vela_app_status.md)	 - get status of an application
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
