+++ documentation/cli/vela_app_ls.md
@@ -33,4 +33,4 @@ vela app ls
 
 * [vela app](vela_app.md)	 - Manage applications
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
