+++ documentation/cli/vela_app_status.md
@@ -32,4 +32,4 @@ vela status <APPLICATION-NAME>
 
 * [vela app](vela_app.md)	 - Manage applications
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
