+++ documentation/cli/vela_cap.md
@@ -26,4 +26,4 @@ Capability Management with config, list, add, remove capabilities
 * [vela cap ls](vela_cap_ls.md)	 - List all capabilities in center
 * [vela cap remove](vela_cap_remove.md)	 - Remove capability from cluster
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
