+++ documentation/cli/vela_cap_add.md
@@ -33,4 +33,4 @@ vela cap add mycenter/route
 
 * [vela cap](vela_cap.md)	 - Capability Management
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
