+++ documentation/cli/vela_cap_center_config.md
@@ -33,4 +33,4 @@ vela cap center config mycenter https://github.com/oam-dev/catalog/cap-center
 
 * [vela cap center](vela_cap_center.md)	 - Manage Capability Center
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
