+++ documentation/cli/vela_cap_center_ls.md
@@ -32,4 +32,4 @@ vela cap center ls
 
 * [vela cap center](vela_cap_center.md)	 - Manage Capability Center
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
