+++ documentation/cli/vela_cap_ls.md
@@ -32,4 +32,4 @@ vela cap ls
 
 * [vela cap](vela_cap.md)	 - Capability Management
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
