+++ documentation/cli/vela_comp.md
@@ -23,9 +23,9 @@ Manage Components
 
 * [vela](vela.md)	 - 
 * [vela comp delete](vela_comp_delete.md)	 - Delete Component From Application
-* [vela comp ls](vela_comp_ls.md)	 - List applications
-* [vela comp deploy](vela_comp_run.md)	 - Init and Run workloads
+* [vela comp deploy](vela_comp_deploy.md)	 - Init and Run workloads
+* [vela comp ls](vela_comp_ls.md)	 - List components
 * [vela comp show](vela_comp_show.md)	 - get component detail
 * [vela comp status](vela_comp_status.md)	 - get status of a component
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
