+++ documentation/cli/vela_comp_delete.md
@@ -33,4 +33,4 @@ vela comp delete frontend
 
 * [vela comp](vela_comp.md)	 - Manage Components
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
