+++ documentation/cli/vela_comp_deploy.md
@@ -19,7 +19,7 @@ vela comp deploy -t <workload-type>
 ### Options
 
 ```
-  -h, --help          help for run
+  -h, --help          help for deploy
   -s, --staging       only save changes locally without real update application
   -t, --type string   specify workload type for application
 ```
@@ -35,4 +35,4 @@ vela comp deploy -t <workload-type>
 
 * [vela comp](vela_comp.md)	 - Manage Components
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
