+++ documentation/cli/vela_comp_ls.md
@@ -1,10 +1,10 @@
 ## vela comp ls
 
-List applications
+List components
 
 ### Synopsis
 
-List applications with workloads, traits, status and created time
+List components with its application, workloads, traits, status and created time
 
 ```
 vela comp ls
@@ -33,4 +33,4 @@ vela comp ls
 
 * [vela comp](vela_comp.md)	 - Manage Components
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
