+++ documentation/cli/vela_comp_show.md
@@ -33,4 +33,4 @@ vela comp show <COMPONENT-NAME>
 
 * [vela comp](vela_comp.md)	 - Manage Components
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
