+++ documentation/cli/vela_completion_bash.md
@@ -36,4 +36,4 @@ vela completion bash
 
 * [vela completion](vela_completion.md)	 - Output shell completion code for the specified shell (bash or zsh)
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
