+++ documentation/cli/vela_completion_zsh.md
@@ -33,4 +33,4 @@ vela completion zsh
 
 * [vela completion](vela_completion.md)	 - Output shell completion code for the specified shell (bash or zsh)
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
