+++ documentation/cli/vela_dashboard.md
@@ -38,4 +38,4 @@ dashboard
 
 * [vela](vela.md)	 - 
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
