+++ documentation/cli/vela_env.md
@@ -1,4 +1,4 @@
-## vela environment
+## vela env
 
 Manage application environments
 
@@ -9,7 +9,7 @@ Manage application environments
 ### Options
 
 ```
-  -h, --help   help for environment
+  -h, --help   help for env
 ```
 
 ### Options inherited from parent commands
@@ -26,4 +26,4 @@ Manage application environments
 * [vela env ls](vela_env_ls.md)	 - List environments
 * [vela env set](vela_env_set.md)	 - Set an environment
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
