+++ documentation/cli/vela_env_init.md
@@ -13,14 +13,16 @@ vela env init <envName>
 ### Examples
 
 ```
-vela env init test --namespace test
+vela env init test --namespace test --email my@email.com
 ```
 
 ### Options
 
 ```
+      --domain string      specify domain your applications
+      --email string       specify email for production TLS Certificate notification
   -h, --help               help for init
-      --namespace string   specify K8s namespace for environment (default "default")
+      --namespace string   specify K8s namespace for env (default "default")
 ```
 
 ### Options inherited from parent commands
@@ -33,4 +35,4 @@ vela env init test --namespace test
 
 * [vela env](vela_env.md)	 - Manage application environments
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
