+++ documentation/cli/vela_env_ls.md
@@ -32,4 +32,4 @@ vela env list [env-name]
 
 * [vela env](vela_env.md)	 - Manage application environments
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
