+++ documentation/cli/vela_install.md
@@ -13,8 +13,11 @@ vela install [flags]
 ### Options
 
 ```
-  -h, --help                     help for install
-  -p, --vela-chart-path string   path to vela core chart to override default chart
+  -h, --help                       help for install
+      --image-pull-policy string   vela core image pull policy, this will align to chart value image.pullPolicy (default "Always")
+      --image-repo string          vela core image repo, this will align to chart value image.repo (default "oamdev/vela-core")
+      --image-tag string           vela core image repo, this will align to chart value image.tag (default "latest")
+  -p, --vela-chart-path string     path to vela core chart to override default chart
 ```
 
 ### Options inherited from parent commands
@@ -27,4 +30,4 @@ vela install [flags]
 
 * [vela](vela.md)	 - 
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
