--- documentation/cli/vela_manualscaler.md
@@ -1,39 +1,0 @@
-## vela manualscaler
-
-Attach manualscaler trait to an app
-
-### Synopsis
-
-Attach manualscaler trait to an app
-
-```
-vela manualscaler <appname> [args]
-```
-
-### Examples
-
-```
-vela manualscaler frontend
-```
-
-### Options
-
-```
-  -a, --app string    create or add into an existing application group
-      --detach        detach trait from component
-  -h, --help          help for manualscaler
-  -r, --replica int    (default 2)
-  -s, --staging       only save changes locally without real update application
-```
-
-### Options inherited from parent commands
-
-```
-  -e, --env string   specify environment name for application
-```
-
-### SEE ALSO
-
-* [vela](vela.md)	 - 
-
-###### Auto generated by spf13/cobra on 11-Sep-2020
