+++ documentation/cli/vela_metrics.md
@@ -1,0 +1,43 @@
+## vela metrics
+
+Attach metrics trait to an app
+
+### Synopsis
+
+Attach metrics trait to an app
+
+```
+vela metrics <appname> [args]
+```
+
+### Examples
+
+```
+vela metrics frontend
+```
+
+### Options
+
+```
+  -a, --app string      create or add into an existing application group
+      --detach          detach trait from component
+      --enabled          (default true)
+  -f, --format string   format of the metrics, default as prometheus (default "prometheus")
+  -h, --help            help for metrics
+      --path string      (default "/metrics")
+      --port int         (default 8080)
+      --scheme string    (default "http")
+  -s, --staging         only save changes locally without real update application
+```
+
+### Options inherited from parent commands
+
+```
+  -e, --env string   specify environment name for application
+```
+
+### SEE ALSO
+
+* [vela](vela.md)	 - 
+
+###### Auto generated by spf13/cobra on 20-Oct-2020
