+++ documentation/cli/vela_route.md
@@ -1,0 +1,42 @@
+## vela route
+
+Attach route trait to an app
+
+### Synopsis
+
+Attach route trait to an app
+
+```
+vela route <appname> [args]
+```
+
+### Examples
+
+```
+vela route frontend
+```
+
+### Options
+
+```
+  -a, --app string      create or add into an existing application group
+      --detach          detach trait from component
+      --domain string   
+  -h, --help            help for route
+      --issuer string   
+      --path string     
+      --port int         (default 443)
+  -s, --staging         only save changes locally without real update application
+```
+
+### Options inherited from parent commands
+
+```
+  -e, --env string   specify environment name for application
+```
+
+### SEE ALSO
+
+* [vela](vela.md)	 - 
+
+###### Auto generated by spf13/cobra on 20-Oct-2020
