+++ documentation/cli/vela_scale.md
@@ -36,4 +36,4 @@ vela scale frontend
 
 * [vela](vela.md)	 - 
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
