+++ documentation/cli/vela_system.md
@@ -24,4 +24,4 @@ system management utilities
 * [vela system info](vela_system_info.md)	 - show vela client and cluster chartPath
 * [vela system update](vela_system_update.md)	 - Sync definition from cluster
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
