+++ documentation/cli/vela_system_info.md
@@ -26,4 +26,4 @@ vela system info [flags]
 
 * [vela system](vela_system.md)	 - system management utilities
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
