+++ documentation/cli/vela_system_update.md
@@ -32,4 +32,4 @@ vela system update
 
 * [vela system](vela_system.md)	 - system management utilities
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
