+++ documentation/cli/vela_template.md
@@ -1,19 +1,15 @@
-## vela logs
+## vela template
 
-Tail pods logs of an application
+Manage templates
 
 ### Synopsis
 
-Tail pods logs of an application
-
-```
-vela logs <appname> [flags]
-```
+Manage templates
 
 ### Options
 
 ```
-  -h, --help   help for logs
+  -h, --help   help for template
 ```
 
 ### Options inherited from parent commands
@@ -25,5 +21,6 @@ vela logs <appname> [flags]
 ### SEE ALSO
 
 * [vela](vela.md)	 - 
+* [vela template context](vela_template_context.md)	 - show context parameters
 
 ###### Auto generated by spf13/cobra on 20-Oct-2020
