+++ documentation/cli/vela_template_context.md
@@ -1,19 +1,25 @@
-## vela logs
+## vela template context
 
-Tail pods logs of an application
+show context parameters
 
 ### Synopsis
 
-Tail pods logs of an application
+show context parameter
 
 ```
-vela logs <appname> [flags]
+vela template context
+```
+
+### Examples
+
+```
+vela template context
 ```
 
 ### Options
 
 ```
-  -h, --help   help for logs
+  -h, --help   help for context
 ```
 
 ### Options inherited from parent commands
@@ -24,6 +30,6 @@ vela logs <appname> [flags]
 
 ### SEE ALSO
 
-* [vela](vela.md)	 - 
+* [vela template](vela_template.md)	 - Manage templates
 
 ###### Auto generated by spf13/cobra on 20-Oct-2020
