+++ documentation/cli/vela_traits.md
@@ -33,4 +33,4 @@ vela traits
 
 * [vela](vela.md)	 - 
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
