+++ documentation/cli/vela_version.md
@@ -26,4 +26,4 @@ vela version [flags]
 
 * [vela](vela.md)	 - 
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
