+++ documentation/cli/vela_workloads.md
@@ -32,4 +32,4 @@ vela workloads
 
 * [vela](vela.md)	 - 
 
-###### Auto generated by spf13/cobra on 11-Sep-2020
+###### Auto generated by spf13/cobra on 20-Oct-2020
